Moving test files to a folder for structure
